% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{\%''\%}
\alias{\%''\%}
\title{Cbind operator}
\usage{
x \%''\% y
}
\arguments{
\item{x}{a vector, mtx, or data frame}

\item{y}{a vector, mtx, or data frame}
}
\value{
cbind ( x,y)
}
\description{
shorthand for cbind
}
\details{
Cbind as usual but using an infix operator
}
