% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammock.R
\name{Csv}
\alias{Csv}
\title{Write a csv file in a convenient way}
\usage{
Csv(x)
}
\arguments{
\item{x}{file name}
}
\description{
A  wrapper around write.csv: you just need to tell it
the object name, file name is made from it by adding the appropriate extension.
}
