% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammock.R
\name{Curry}
\alias{Curry}
\title{Partial function application}
\usage{
Curry(FUN, ...)
}
\arguments{
\item{FUN}{a function}

\item{...}{dot dot dot}
}
\value{
a function
}
\description{
Partial function application using do.call
by Byron Ellis, https://stat.ethz.ch/pipermail/r-devel/2007-November/047318.html
But see also discussion in R-devel list (Hadley Wickham, Luke Tierney et al)
}
