% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Df.R
\name{Df}
\alias{Df}
\alias{$<-.Df}
\alias{[[<-.Df}
\title{Nice data frame}
\usage{
Df(
  ...,
  vars = NULL,
  info = NULL,
  row.names = NULL,
  check.rows = FALSE,
  check.names = FALSE,
  stringsAsFactors = FALSE,
  emptyStringsAsNA = TRUE
)

\method{$}{Df}(x, name) <- value

\method{[[}{Df}(x, i, j) <- value
}
\arguments{
\item{...}{these arguments are of either the form value or tag = value. Component names are created based on the tag (if present) or the deparsed argument itself.}

\item{vars}{A data.frame with meta-data on variables}

\item{info}{An informative title of the data frame}

\item{row.names}{NULL or a single integer or character string specifying a column to be used as row names, or a character or integer vector giving the row names for the data frame.}

\item{check.rows}{if TRUE then the rows are checked for consistency of length and names.}

\item{check.names}{logical; defaults to FALSE. If TRUE then the names of the variables in the data frame are checked to ensure that they are syntactically valid variable names and are not duplicated. If necessary they are adjusted (by make.names) so that they are.}

\item{stringsAsFactors}{logical: should character vectors be converted to factors? Defaults (of course!!) to FALSE.}

\item{emptyStringsAsNA}{logical: should empty strings (within character vectors) be converted to NA? Defaults to TRUE.}
}
\value{
a data frame

\code{NULL}

\code{NULL}
}
\description{
A version of data.frame with stringsAsFactors and check.names defaulting to FALSE.
}
\seealso{
\code{\link{data.frame}}
}
