% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab}
\alias{grab}
\alias{grabn}
\alias{grab.list}
\alias{grab.data.frame}
\alias{grab.character}
\title{Select columns from a data frame using regular expressions}
\usage{
grab(x, pattern, ...)

\method{grab}{list}(x, pattern, ...)

\method{grab}{data.frame}(x, pattern, ...)

\method{grab}{character}(x, pattern, ...)
}
\arguments{
\item{x}{an object}

\item{pattern}{a pattern}

\item{...}{...}
}
\value{
something

\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Returns the set of columns in a data frame whose names
correspond to a specific pattern
}
