% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammock.R
\docType{package}
\name{hammock-package}
\alias{hammock-package}
\alias{hammock}
\title{Hammock package}
\description{
\tabular{ll}{
Package: \tab hammock\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-08-02\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\details{
Functions for lazy people

\code{\link{Load}} loads \code{file} and returns a value 
rather than creating the objects in Global Env.
}
\examples{
NULL


}
\references{
\url{http://en.wikipedia.org/wiki/hammock}
}
\seealso{
\code{\link{Load}}
}
\author{
Kenn Konstabel \email{kenn.konstabel AT tai.ee}
}
\keyword{package}
