% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Df-methods.R
\name{labs}
\alias{labs}
\alias{info}
\alias{labs.Df}
\alias{labs<-}
\alias{labs<-.Df}
\alias{info.Df}
\alias{info<-}
\alias{info<-.Df}
\title{Get or set lengend and info on a Df}
\usage{
labs(x)

\method{labs}{Df}(x)

labs(x) <- value

\method{labs}{Df}(x) <- value

info(x)

\method{labs}{Df}(x)

info(x) <- value

\method{"info"}{Df}(x) <- value
}
\arguments{
\item{x}{an object}

\item{value}{a value}
}
\value{
Invisible x

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Sets or gets attributes .... ....
}
