% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List.R
\name{List}
\alias{List}
\title{Make a named list}
\usage{
List(..., ..umn = TRUE)
}
\arguments{
\item{...}{elements of list}

\item{..umn}{use make.names?}
}
\value{
a named list
}
\description{
Make a named list. List(a,b,c) is equivalent to list(a=a, b=b, c=c);
List(o=a, b) is equivalent to list(o=a, b=b). Unnamed list elements are used
to create names: either using make.names or force.
}
