% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{Load}
\alias{Load}
\title{Load save'd objects and return them as a list}
\usage{
Load(
  file,
  to.list = TRUE,
  Unlist = TRUE,
  spray = FALSE,
  delete.parent = FALSE,
  Attach = FALSE
)
}
\arguments{
\item{file}{the file to load}

\item{to.list}{if TRUE, a list is returned, otherwise an environment}

\item{Unlist}{whether to "unlist" length 1 lists (default is TRUE)}

\item{spray}{whether to create these objects in the global environment (default is FALSE)}

\item{delete.parent}{if an environment is returned, whether to set its parent env to an empty environment (to make a smaller object)}

\item{Attach}{to attach the loaded objects to the search path}
}
\value{
a list or environment (depending on to.list) or other object (if there is just one object saved and Unlist equals TRUE) or invisibly NULL if Attach==TRUE
}
\description{
Obects in `file` are loaded and returned as a value (instead of 
creating these objects in the global environment). The value can be
a list (default) or an environment (if to.list is FALSE); if there
is just one object in the file to be read, it is by default unlisted
and just the objects' value is returned.
}
\details{
This is useful to avoid overwriting objects in your workspace, 
for using only parts of a file, and inspecting the contents of files 
before you load them. You can set your own names to the objects instead
of using the ones from the file. E.g., 

\code{new.name <- Load("manyobjects.rda")$old.name }
\code{ls(Load("manyobjects.rda"))}
}
\seealso{
\code{\link{load}}
}
