% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_sources.R
\name{ls.sources}
\alias{ls.sources}
\title{List objects in R soucre files.}
\usage{
ls.sources(DIR = ".", use.basename = TRUE)
}
\arguments{
\item{DIR}{name of the directory, defaults to "."}

\item{use.basename}{Whether to use basename (TRUE) or full name for list names}
}
\value{
a list with one component per each *.R file in the folder
}
\description{
ls.soucres.
}
\details{
The *.R files in a folder (defaults to the current working
directory) are Source'd and the names of the objects listed.
This does not add anything to the global environment (see
`Source`).

This is useful to get an overview of what's where e.g in 
an R/ subfolder of a package.
}
\seealso{
\code{\link{Source}}
}
