% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammock.R
\name{nload}
\alias{nload}
\title{Source and attach a set of functions from a file as a pseudo package}
\usage{
nload(what, www = FALSE, DIR = "Rfns")
}
\arguments{
\item{what}{a name}

\item{www}{read from www?}

\item{DIR}{name of the folder where the files are located}
}
\value{
nothing useful
}
\description{
A set of functions or other objects is read in and attached
to the search path. This will be reworked: adding just an
xtra parameter to Source.
}
