% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open.R
\name{Open}
\alias{Open}
\title{Open a file using windows or linux default file associations}
\usage{
Open(x = "")
}
\arguments{
\item{x}{the file to load}
}
\value{
NULL; used for side effect
}
\description{
Open a file using windows default file associations (call with 
no arguments to open file explorer in the current working directory)
Don't know how to do it on Mac.
}
