% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_src.R
\name{parse_source}
\alias{parse_source}
\title{Parse source}
\usage{
parse_source(file, keep = "all", browse = FALSE)
}
\arguments{
\item{file}{a file name}

\item{keep}{either 'all' or 'fun'}
}
\value{
value**
}
\description{
Parses an R source file to find objects (experimental)
}
\details{
Parse source
}
