% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_src.R
\name{parse_sources}
\alias{parse_sources}
\title{Parse sources}
\usage{
parse_sources(DIR = ".", SORT = TRUE)
}
\arguments{
\item{DIR}{a file name}

\item{SORT}{boolean}
}
\value{
value**asd functions
}
\description{
Parse sources
}
