% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{\%=\%}
\alias{\%=\%}
\title{Rbind operator}
\usage{
x \%=\% y
}
\arguments{
\item{x}{a vector, mtx, or data frame}

\item{y}{a vector, mtx, or data frame}
}
\value{
rbind ( x,y)
}
\description{
shorthand for rbind
}
\details{
Rbind as usual but using an infix operator
}
