% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Spss.R
\name{read.Spss}
\alias{read.Spss}
\title{Read.spss to Df}
\usage{
read.Spss(x, info = "", usv = FALSE, browse = FALSE, ...)
}
\arguments{
\item{x}{a file name}

\item{usv}{use value labels? passed to read.spss}

\item{browse}{FALSE}

\item{...}{additional parameters passed to read.spss}
}
\value{
a Df
}
\description{
SPSS to Df
}
