% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{\%R\%}
\alias{\%R\%}
\title{Rounding operator}
\usage{
x \%R\% digits
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{a number}
}
\value{
Rounded value
}
\description{
Rounds the values in its first argument to the specified number of decimal places
}
\details{
Rounding as usual but using an infix operator
}
