% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save.R
\name{Save}
\alias{Save}
\title{Save-as-you-go}
\usage{
Save(..., file)
}
\arguments{
\item{...}{objs (should be named - othewise funny manes will be used)}

\item{file}{filename - must be there!}
}
\value{
nothing useful ?
}
\description{
Saves any objects, not just named ones
}
\details{
Save

\code{Save(bla = matrix(sample(1:10),2,4), foo = A[1:5, 2:7]) }
}
\seealso{
\code{\link{load}}
}
