% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tags.R
\name{show_tags}
\alias{show_tags}
\title{Show tags of a Df in lower-right part of Rstudio screen}
\usage{
show_tags(
  x,
  cols = NULL,
  NW = "30\%",
  TW = "100\%",
  filename = NULL,
  showit = TRUE
)
}
\arguments{
\item{x}{a Df}

\item{cols}{Column numbers to show}

\item{NW}{width of leftmost column}

\item{filename}{file name}

\item{showit}{TRUE or FALSE}
}
\value{
table
}
\description{
Table of tags
}
