% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammock.R
\name{Smerge}
\alias{Smerge}
\title{Simple merging of data frames}
\usage{
Smerge(x, y, by, finally = data.frame, unrowname = TRUE, ...)
}
\arguments{
\item{x}{a data frame}

\item{y}{a data frame}

\item{by}{character: name of the variable to be used for merging}

\item{finally}{function to be applied to the two data frames after they're made to match}

\item{unrowname}{logical: delete row names?}

\item{...}{parameters to be passed to `finally`}
}
\value{
the 2 data frames made to match row by row, combined as specified by `finally`
}
\description{
Merges two data frame in a simple and safe way. 
The result will have all the rows from the first data frame 
exactly in the same order. The new rows from the second data
frame are added to the end.
}
