% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammock.R
\name{sort.data.frame}
\alias{sort.data.frame}
\title{Sort a data frame using one of its variabes as index}
\usage{
\method{sort}{data.frame}(x, decreasing = FALSE, by, ...)
}
\arguments{
\item{x}{a data frame}

\item{decreasing}{passed to order}

\item{by}{name of the}

\item{...}{additional args passed to next method if they get a chance}
}
\value{
a sorted data frame
}
\description{
Sorts a data frame using a variable within that data frame as an index.
it's basically just x[order(x[[by]])]
Could be written as a method of `sort` for data frame but need 2 figure out how to do it
}
