% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source.R
\name{Source}
\alias{Source}
\title{Source a file and return all the created objects as a list or environment.}
\usage{
Source(
  file,
  to.list = TRUE,
  Unlist = TRUE,
  spray = FALSE,
  delete.parent = FALSE,
  Attach = FALSE,
  First = list(),
  Setwd = NULL
)
}
\arguments{
\item{file}{the file to load}

\item{to.list}{if TRUE, a list is returned, otherwise an environment}

\item{Unlist}{whether to "unlist" length 1 lists (default is TRUE)}

\item{spray}{whether to create these objects in the global environment (default is FALSE)}

\item{delete.parent}{if an environment is returned, whether to set its parent env to an empty environment (to make a smaller object)}

\item{Attach}{to attach the Source'd environment}

\item{First}{a list of values to be included in the output (that the Source'd script can possibly use)}
}
\value{
a list or environment (depending on to.list) or other object (if there is just one object saved and Unlist equals TRUE)
}
\description{
The parameters are analogous to Load.
}
\details{
The `file` is sourced but evaluated in a special environment, and
the contents of that environment are then returned as a list or
environment. If just one object is created in the sourced code, 
then by default just that object's value is returned (if Unlist is TRUE).

This is useful to avoid overwriting objects in your workspace, 
for using only parts of a file, and inspecting the contents of files 
before you load them. You can set your own names to the objects instead
of using the ones from the file. E.g., new.name <- Load("manyobjects.rda")$old.name,
or ls(Load("manyobjects.rda"))
}
\seealso{
\code{\link{source}}
}
