% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.R
\name{strip}
\alias{strip}
\title{Make a function to return its own environment}
\usage{
strip(fun.)
}
\arguments{
\item{fun.}{a function}
}
\value{
a modified function
}
\description{
Modifies a function so that it returns its own environment.
The new function returns an environment with all intermediate
variables calculated in the original function. The function's
value is assigned to `result..`
Doesn't work with S3 generics and primitive.
}
