% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xapply.R
\name{xapply}
\alias{xapply}
\title{Apply an expression over a list of "arguments"}
\usage{
xapply(., ..., SIMPLIFY = TRUE, FUN = cbind)
}
\arguments{
\item{.}{an expression}

\item{...}{named args (must be named!)}

\item{SIMPLIFY}{to return a simple vector or matrix if possible?}

\item{FUN}{function to be called to simplify the result (e.g rbind, cbind, data.frame)}
}
\value{
a value
}
\description{
"applies" an expression (.) to a list of "arguments". 
Analogous to mapply but works with expressions rather than functions.\cr
Examples:\cr
mxapply(a+b, a=1:5, b=2:6)\cr
mxapply(c(a+b, a-b), a=1:5, b=2:6)\cr
mxapply(c(a+b, a-b), a=1:5, b=2:6, FUN=rbind)\cr
mxapply(c(a+b, a-b), a=1:5, b=2:6, FUN=data.frame)\cr
}
